Add files via upload
In the world, Alzheimer's disease (AD), a neurodegenerative condition, is the most prevalent 
cause of dementia and the sixth greatest cause of mortality. Memory loss that worsens with 
time, cognitive impairment, and behavioural abnormalities are the disease's hallmarks. The 
present therapies for AD simply alleviate symptoms; there is no known cure for the 
condition.
This study identified risk factors for AD using descriptive statistics, clustering, and logistic 
regression, which is presented in this report. An analysis of 317 individuals' data revealed 
that older age, male gender, lower socioeconomic level, lower educational attainment, and 
lower cognitive test scores were all linked to a higher risk of AD.
According to the study's findings, there are a number of AD risk factors that can be changed. 
A reduction in the prevalence of AD and an improvement in the quality of life for those who 
have the illness are both possible outcomes of interventions that focus on these risk factors.